+++ hspt-survey/src/main/java/org/hspt/entity/request/ReqSurveyType.java
@@ -20,7 +20,7 @@ public class ReqSurveyType extends BaseRequest {
 
     @ApiModelProperty("问卷类型编号")
     @NotNull
-    private Integer pkSurveyType;
+    private long pkSurveyType;
 
     @NotBlank
     @ApiModelProperty("问卷类型名称")
